requirements.txt
google-generativeai==0.8.5
flask==3.0.0
python-dotenv==1.0.0 